% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_tools.R
\name{align_to}
\alias{align_to}
\title{Align a SpatRaster or sf object to a template CRS or geometry}
\usage{
align_to(src, template, categorical = TRUE, pixel_size = NULL)
}
\arguments{
\item{src}{A `SpatRaster` or `sf` object to be projected / aligned.}

\item{template}{Either a `SpatRaster` whose geometry should be matched,
or an `sf` object whose CRS should be matched.}

\item{categorical}{Logical. If `TRUE`, use nearest-neighbor resampling;
if `FALSE`, use bilinear interpolation.}

\item{pixel_size}{If `src` is `sf`, this sets the raster resolution (in map units).
If NULL and `template` is a `SpatRaster`, resolution is inferred from template.
Default is `NULL`.}
}
\value{
A `SpatRaster` aligned to the template.
}
\description{
Projects a `SpatRaster` or `sf` object (`src`) to match either the full geometry
(CRS, extent, resolution) of a `SpatRaster` template, or only the CRS
of an `sf` template.
}
