% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{apply_scale}
\alias{apply_scale}
\title{Apply Value Rescaling to a List of SpatRasters}
\usage{
apply_scale(rlist, new_range, cat = FALSE, quiet = TRUE)
}
\arguments{
\item{rlist}{A list of `terra::SpatRaster` objects to be rescaled.}

\item{new_range}{A numeric vector of length 2 defining the new value range
for rescaling (`c(new_min, new_max)`). If `NULL`, no rescaling is applied.}

\item{cat}{Logical. If `TRUE`, any raster that is internally stored as an
integer type (`terra::is.int()`) will be rounded after scaling.}

\item{quiet}{Logical. If `TRUE`, suppresses progress messages.}
}
\value{
A list of `terra::SpatRaster` objects, rescaled where appropriate.
}
\description{
Applies linear rescaling to each element of a list of `SpatRaster` objects,
using a user-defined numeric range. Optionally rounds integer rasters after
scaling.
}
\details{
For each raster:
* Its current min/max is compared with `new_range`.
* If identical, no rescaling is performed.
* Otherwise, the raster is passed to `scale_vars()`.

The function preserves the structure of the input list.
}
\examples{
\dontrun{
library(terra)
r1 <- rast(matrix(runif(100), 10, 10))
r2 <- rast(matrix(runif(100), 10, 10))
lst <- list(a = r1, b = r2)

scaled <- apply_scale(lst, c(0, 1))
}

}
