% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AssayTab-class}
\alias{AssayTab-class}
\alias{AssayTab}
\alias{MSAssayTab,}
\alias{MicroarrayAssayTab,}
\alias{SeqAssayTab,}
\alias{NMRAssayTab}
\alias{MSAssayTab-class}
\alias{MSAssayTab}
\alias{MicroarrayAssayTab-class}
\alias{MicroarrayAssayTab}
\alias{SeqAssayTab-class}
\alias{SeqAssayTab}
\alias{NMRAssayTab-class}
\title{S4 Class AssayTab}
\description{
An S4 class to store information of an AssayTab.
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-Tab
dataset.}

\item{\code{study.filename}}{A character vector with the filename of the study to
which this assay belongs to.}

\item{\code{study.identifier}}{A character vector with the identifier of the study to
which this assay belongs to.}

\item{\code{assay.filename}}{A character vector with the filename of the assay.}

\item{\code{assay.file}}{A data frame object with the contents of the assay file.}

\item{\code{assay.technology.type}}{A character vector with the technology type of the
assay.}

\item{\code{assay.measurement.type}}{A character vector with the measurement type of
the assay.}

\item{\code{assay.names}}{A data frame object with the assay names for the assay file.}

\item{\code{data.filenames}}{A data frame object with the data filenames for the assay.}
}}

\examples{
showClass("AssayTab")
showClass("MSAssayTab")
showClass("MicroarrayAssayTab")
showClass("SeqAssayTab")
showClass("NMRAssayTab")

}
\seealso{
\code{\link{ISATab-class}}, \code{\link{ISATab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
\keyword{classes}
