% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_tools.R
\name{convert_to_decimal_degrees}
\alias{convert_to_decimal_degrees}
\title{Convert spatial object to decimal degrees (EPSG:4326)}
\usage{
convert_to_decimal_degrees(obj)
}
\arguments{
\item{obj}{A `SpatRaster` or `sf` object with defined CRS.}
}
\value{
Transformed object in EPSG:4326.
}
\description{
Projects a `SpatRaster` or `sf` object into geographic coordinates.
}
\examples{
# Create test data
df <- data.frame(long = c(277000,389000,389000,611000),
                 lat = c(442000,442000,221000,221000))
sf_obj <- st_as_sf(df, coords = names(df), crs = 32631, remove = FALSE)

# Convert to decimal degrees
convert_to_decimal_degrees(sf_obj)
}
