% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_reshape.R
\name{criteria_reshape}
\alias{criteria_reshape}
\title{#Load example data
path <- system.file("extdata", "multi_species_criteria.csv", package = "risa")
df <- read.csv(path)}
\usage{
criteria_reshape(x)
}
\description{
#Inspect dataframe
head(df)
}
\details{
#Reshape criteria table
crit_list <- criteria_reshape(df)
crit_list
}
