% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{df_to_list}
\alias{df_to_list}
\title{Split a data.frame or sf into a list of sf by group}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{A data.frame or `sf` with at least two coordinate columns and optional group column.}
}
\value{
A named list of `sf` objects, one per group level.
}
\description{
Splits data.frame rows into separate `sf` objects based on a grouping factor (column 3).
}
\examples{
# Create test data
df <- data.frame(long = c(1,2,2,4,2,5,4,6),
                 lat = c(4,4,2,2,4,4,2,2),
                 species=rep(c("sp1", "sp2"), each=4))

#create vector list
vec_list <- df_to_list(df)
vec_list
}
