% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{[,ISATab,character,ANY,ANY-method}
\alias{[,ISATab,character,ANY,ANY-method}
\alias{[<-,ISATab,character,ANY,ANY-method}
\alias{[,AssayTab,character,ANY,ANY-method}
\alias{[<-,AssayTab,character,ANY,ANY-method}
\title{Extract and replace methods}
\usage{
\S4method{[}{ISATab,character,ANY,ANY}(x, i)

\S4method{[}{ISATab,character,ANY,ANY}(x, i) <- value

\S4method{[}{AssayTab,character,ANY,ANY}(x, i)

\S4method{[}{AssayTab,character,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{S4 object of class \code{ISTab} or \code{AssayTab}}

\item{i}{character, name of a slot to extract or set}

\item{value}{ANY, value for the slotname specified by i to be replaced}
}
\description{
Extract or replace slots from an object of class \linkS4class{ISATab}.

Extract or replace slots from an object of class \linkS4class{AssayTab}.
}
