% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{getAnnotatedDataFrameAssay}
\alias{getAnnotatedDataFrameAssay}
\title{Get an Annotated Data Frame from an Assay file.}
\usage{
getAnnotatedDataFrameAssay(isa, assay.filename)
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}.}}

\item{assay.filename}{A character vector with the filename of an assay.}
}
\value{
An annotated data frame.
}
\description{
Retrieves an Annotated Data Frame from an Assay file.
}
\examples{
faahkoISAzip <- system.file("extdata", "faahKO-metadata.zip", package = "Risa")
faahkoISA <- readISAtab(path = file.path(system.file("extdata", package = "Risa"),
                                         "faahkoISA"),
                        zipfile = faahkoISAzip,
                         verbose =TRUE)
faahkoADF <- getAnnotatedDataFrameAssay(faahkoISA, faahkoISA@assay.filenames[[1]])
str(faahkoADF)

armstrongISAzip <- system.file("extdata", "ARMSTRONG-S-3-metadata.zip", package = "Risa")
armstrongISA <- readISAtab(path = file.path(system.file("extdata",
                                                        package = "Risa"),
                                            "armstrongISA"),
                           zipfile = armstrongISAzip,
                           verbose = TRUE)
armstrongADF <- getAnnotatedDataFrameAssay(isa = armstrongISA,
                                           assay.filename = armstrongISA@assay.filenames[[1]])
str(armstrongADF)

}
\seealso{
\code{\link{ISATab-class}}, \code{\link{AnnotatedDataFrame}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
