% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getAssayRawDataFilenames}
\alias{getAssayRawDataFilenames}
\alias{getAssayRawDataFilenames,AssayTab,logical-method}
\alias{getAssayRawDataFilenames,MSAssayTab,logical-method}
\alias{getAssayRawDataFilenames,MicroarrayAssayTab,logical-method}
\alias{getAssayRawDataFilenames,SeqAssayTab,logical-method}
\alias{getAssayRawDataFilenames,NMRAssayTab,logical-method}
\title{getAssayRawDataFilenames Generic}
\usage{
getAssayRawDataFilenames(.Object, full.path)

\S4method{getAssayRawDataFilenames}{AssayTab,logical}(.Object,
  full.path = TRUE)

\S4method{getAssayRawDataFilenames}{MSAssayTab,logical}(.Object,
  full.path = TRUE)

\S4method{getAssayRawDataFilenames}{MicroarrayAssayTab,logical}(.Object,
  full.path = TRUE)

\S4method{getAssayRawDataFilenames}{SeqAssayTab,logical}(.Object,
  full.path = TRUE)

\S4method{getAssayRawDataFilenames}{NMRAssayTab,logical}(.Object,
  full.path = TRUE)
}
\arguments{
\item{.Object}{An object of a defined S4 class (\linkS4class{AssayTab},
\linkS4class{MSAssayTab}, \linkS4class{MicroarrayAssayTab},
\linkS4class{SeqAssayTab}, \linkS4class{NMRAssayTab}).}

\item{full.path}{a length-one logical vector indicating whether the filenames
should be retrieved with their full path [default: TRUE],
or not (FALSE).}
}
\description{
Generic function to retrieve the raw data files for a particular assay type.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = AssayTab,full.path = logical}: Retrieves the raw data filenames for a generic assay.

\item \code{.Object = MSAssayTab,full.path = logical}: Retrieves the raw data filenames for an assay whose technology type is mass spectrometry. These data filenames correspond to those specified in the column 'Raw Spectral Data File'.

\item \code{.Object = MicroarrayAssayTab,full.path = logical}: Retrieves the raw data filenames for an assay whose technology type is DNA microarray.

\item \code{.Object = SeqAssayTab,full.path = logical}: Retrieves the raw data filenames for an assay whose technology type is nucleotide sequencing.

\item \code{.Object = NMRAssayTab,full.path = logical}: Retrieves the raw data filenames for an assay whose technology type is NMR spectroscopy.
}}

