% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-microarray.R
\name{getMIAMEMetadata}
\alias{getMIAMEMetadata}
\title{Retrieves the MIAME information about a microarray experiment.}
\usage{
getMIAMEMetadata(isa, assay.filename)
}
\arguments{
\item{isa}{an object of class \linkS4class{ISATab}.}

\item{assay.filename}{A string indicating the name of an assay file. It must
correspond to a microarray-based assay for the function
to return a valid value.}
}
\value{
It retrieves an object from the MIAME class or an error message.
}
\description{
If the assay.filename uses the microarray technology, it retrieves an object
of the class MIAME with the minimum information about the experiment. Otherwise,
it outputs a message indicating that the method is not valid for assays not
using the microarray technology.
}
\examples{
armstrongISAzip <- system.file("extdata", "ARMSTRONG-S-3-metadata.zip",
                               package = "Risa")
armstrongISA <- readISAtab(path = file.path(system.file("extdata",
                                            package = "Risa"),
                                            "armstrongISA"),
                           zipfile = armstrongISAzip,
                           verbose = TRUE)
assay.filename <- armstrongISA@assay.filenames[[1]]
miame <- getMIAMEMetadata(armstrongISA, assay.filename)
str(miame)

}
\seealso{
\code{\link{isMicroarrayAssay}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
