% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-microarray.R
\name{getMicroarrayAssayFilenames}
\alias{getMicroarrayAssayFilenames}
\title{Listing filenames of microarray-based asssay from an ISAtab dataset.}
\usage{
getMicroarrayAssayFilenames(isa)
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}
}
\value{
a character vector with microarray-based assay filenames.
}
\description{
Given an object of class \linkS4class{ISATab}, this method retrieves the list
of asssay filenames in an ISAtab dataset, which correspond to
microarray-based assays.
}
\examples{
armstrongISAzip <- system.file("extdata", "ARMSTRONG-S-3-metadata.zip",
                               package = "Risa")
armstrongISA <- readISAtab(path = file.path(system.file("extdata",
                                                        package = "Risa"),
                                            "armstrongISA"),
                           zipfile = armstrongISAzip,
                           verbose = TRUE)
getMicroarrayAssayFilenames(armstrongISA)

}
\seealso{
\code{\link{isMicroarrayAssay}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
