% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-ms.R
\name{getMSAssayFilenames}
\alias{getMSAssayFilenames}
\title{Retrieves a vector with the assay filenames corresponding to mass spectrometry.}
\usage{
getMSAssayFilenames(isa)
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}
}
\value{
A character vector with the assay filenames corresponding to mass
        spectrometry.
}
\description{
Given an object of class \linkS4class{ISATab}, this method retrieves a vector
with the assay filenames corresponding to mass spectrometry (identified by
having a column called 'Raw Spectral Data File').
}
\examples{
data.dir <- system.file("extdata", package="Risa")
isazip <- system.file("extdata","faahKO-metadata.zip", package="Risa")
faahkoISA <- readISAtab(zipfile = isazip, path = file.path(data.dir,"faahKOISA"), verbose =TRUE)
ms.assay.filenames <- getMSAssayFilenames(faahkoISA)

}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
