% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getRawDataFilenames}
\alias{getRawDataFilenames}
\title{Retrieve the raw data filenames in an ISATab dataset object.}
\usage{
getRawDataFilenames(.Object, full.path = TRUE)
}
\arguments{
\item{.Object}{an object of class \linkS4class{ISATab}}

\item{full.path}{a length-one logical vector indicating whether the filenames
should be retrieved with their full path [default: TRUE],
or not (FALSE).}
}
\value{
list of matrices (one matrix per assay file) containing the raw data filenames
}
\description{
Retrieves the raw data filenames in an ISATab dataset given an object from
the \code{\link{ISATab-class}} and a logical value indicating if the 
filenames are retrieved with their full path or not.
}
