% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{getStudyFilename}
\alias{getStudyFilename}
\title{Retrieves the study filename to which an assay filename corresponds.}
\usage{
getStudyFilename(isa, assay.filename)
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}}

\item{assay.filename}{A character vector with the name of the assay file for
which the study filename needs to be retrieved.}
}
\value{
A character vector containing the study filename.
}
\description{
Retrieve the filename of a study to which a known assay filename
             corresponds.
}
\examples{
faahkoISA <- readISAtab(path = find.package("faahKO"))
getStudyFilename(isa = faahkoISA,
                 assay.filename = faahkoISA@assay.filenames[[1]])
}
\seealso{
\code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
