% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ISATab-class}
\alias{ISATab-class}
\alias{ISATab}
\title{S4 Class ISATab}
\description{
An S4 class to store information from an ISATab dataset, including an
investigation file, one or more study files, and one or more assay files for
each study file.
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-Tab
dataset.}

\item{\code{investigation.filename}}{A length-one character vector containing the
investigation filename (by definition starting
with "i_").}

\item{\code{investigation.file}}{A data frame with the contents of the investigation
file.}

\item{\code{investigation.identifier}}{A length-one character vector containing an
identifier or an accession number for the
investigation provided by a repository (SHOULD
be locally unique).}

\item{\code{study.identifiers}}{A character vector whose elements are the study
identifiers.}

\item{\code{study.titles}}{A character vector whose elements are the study titles.}

\item{\code{study.descriptions}}{A character vector whose elements are the study
descriptions.}

\item{\code{study.contacts}}{A character matrix where each row contains the study
contacts for a specific study. The rowname specifies to
which study the contacts belong.}

\item{\code{study.contacts.affiliations}}{A character matrix where each row contains
the study contacts affiliations for a
specific study. The rowname specifies to
which study the contacts affiliations belong.}

\item{\code{study.filenames}}{A named character vector whose elements are the names of
the study files. The vector element names are specified
as the study identifiers.}

\item{\code{study.files}}{A list object of data frames with the contents of the study
files.}

\item{\code{assay.filenames}}{A character vector whose elements list the names of the 
assay files.}

\item{\code{assay.filenames.per.study}}{A list object of assay filenames for each
study in the investigation.}

\item{\code{assay.files}}{A list object of data frames with the contents of the assay
files.}

\item{\code{assay.files.per.study}}{A list object  of data frames containing assay
files contents for each study in the investigation.}

\item{\code{assay.names}}{A list object of data frames, containing, for each assay file,
the column of the assay file corresponding to the "Assay
Name".}

\item{\code{assay.technology.types}}{A character vector whose elements are the
technology types for each assay.}

\item{\code{assay.technology.types.per.study}}{A list object of assay technology types
for each study in the investigation.}

\item{\code{assay.measurement.types}}{A character vector whose elements list the
measurement types for each assay.}

\item{\code{assay.measurement.types.per.study}}{A list object of assay measurement
types for each study in the investigation.}

\item{\code{data.filenames}}{A list object of data frames, with each data frame
containing the names of the data files for each sample
in an assay.}

\item{\code{samples}}{A character vector whose elements list the names of all samples.}

\item{\code{samples.per.study}}{A list object of character vectors, where the vectors
contain the samples names for each study in the
investigation.}

\item{\code{samples.per.assay.filename}}{A list object of character vectors, where the
vectors contain the samples names for each
assay in the investigation.}

\item{\code{assay.filenames.per.sample}}{A list object with the length of the number
of samples used in an investigation, where
each element is a list object of assay
filenames in which that particular sample
has been used.}

\item{\code{sample.to.rawdatafile}}{A list object of data frames with samples
associated with raw data files.}

\item{\code{sample.to.assayname}}{A list object which maintains the association
between samples and assay names.}

\item{\code{rawdatafile.to.sample}}{A list object which mantains the association
between raw data file and samples.}

\item{\code{assayname.to.sample}}{A list object mantaining the associations between
assay names and samples.}

\item{\code{factors}}{A list object containing R factor objects for each of the studies
(with factor names and factor levels).}

\item{\code{treatments}}{A list object of  data frames with the combination of factor
values, per study file, or a factor object, if there is a
single study.}

\item{\code{groups}}{A list object with one element per study file, and each element
is a list object of the samples for the corresponding treatment.}

\item{\code{assay.tabs}}{A list object, where each element is an object of the class
AssayTab or its specialisations (MSAssayTab or
MicroarrayAssayTab), representing the information of each of
the assay files in the investigation.}
}}

