% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-microarray.R
\name{isMicroarrayAssay}
\alias{isMicroarrayAssay}
\title{Determines if an assay is microarray-based or not.}
\usage{
isMicroarrayAssay(isa, assay.filename)
}
\arguments{
\item{isa}{an object of class \linkS4class{ISATab}.}

\item{assay.filename}{A string indicating the name of an assay file. It must
correspond to a microarray-based assay for the function
to return a valid value.}
}
\value{
Retrieves a boolean value.
}
\description{
Retrieves a boolean value indicating wether the assay filename given as
parameter corresponds to a microarray-based assay or not.
}
\examples{
data.dir <- system.file("extdata", package = "Risa")
isazip <- system.file("extdata", "ARMSTRONG-S-3-metadata.zip",
                      package = "Risa")
isa <- readISAtab(path = file.path(data.dir, "ARMSTRONG_ISA"),
                  zipfile = isazip,
                  verbose = TRUE)
assay.filename <- isa@assay.filenames[[1]]
isMicroarrayAssay(isa, assay.filename)

}
\seealso{
\code{\link{getMIAMEMetadata}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
