% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-ms.R
\name{isMSAssay}
\alias{isMSAssay}
\title{Indicates if an assay filename corresponds to a mass spectrometry assay.}
\usage{
isMSAssay(isa, assay.filename)
}
\arguments{
\item{isa}{an object of class \linkS4class{ISATab}.}

\item{assay.filename}{the filename of an assay file.}
}
\value{
It retrieves a boolean value indicating wether the assay is a mass
        spectrometry assay or not.
}
\description{
Indicates if an assay filename corresponds to a mass spectrometry assay.
}
\examples{
faahkoISA <- readISAtab(find.package("faahKO"))
assay.filename <- faahkoISA["assay.filenames"][[1]]
isMSAssay(faahkoISA, assay.filename)

}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
