% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MicroarrayAssayTab-class}
\alias{MicroarrayAssayTab-class}
\alias{MicroarrayAssayTab}
\title{S4 Class MicroarrayAssayTab}
\description{
An S4 class to store information in an object about an Microarray Assay. It
is an extension of the \linkS4class{AssayTab} class, where the
assay.technology.type is specified as "DNA microarray".
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-Tab
dataset.}

\item{\code{study.filename}}{A character vector with the filename of the study to
which this assay belongs to.}

\item{\code{study.identifier}}{A character vector with the identifier of the study to
which this assay belongs to.}

\item{\code{assay.filename}}{A character vector with the filename of the assay.}

\item{\code{assay.file}}{A data frame object with the contents of the assay file.}

\item{\code{assay.technology.type}}{A character vector specifying the technology type
of the assay as "DNA microarray".}

\item{\code{assay.measurement.type}}{A character vector with the measurement type of
the assay.}

\item{\code{assay.names}}{A data frame object with the assay names for the assay file.}

\item{\code{data.filenames}}{A data frame object with the data filenames for the assay.}
}}

\examples{
showClass("MicroarrayAssayTab")

}
\seealso{
\code{\link{[,AssayTab,character,ANY,ANY-method}}, 
         \code{\link{[<-,AssayTab,character,ANY,ANY-method}},
         \code{\link{AssayTab-class}}, \code{\link{ISATab-class}},
         \code{\link{MSAssayTab-class}}, \code{\link{SeqAssayTab-class}},
         \code{\link{NMRAssayTab-class}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
\keyword{classes}
