% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-ms.R
\name{processAssayXcmsSet.1factor}
\alias{processAssayXcmsSet.1factor}
\title{Build an xcmsSet object given a mass spectrometry assay and considering the
first factor defined}
\usage{
processAssayXcmsSet.1factor(isa, assay.filename, ...)
}
\arguments{
\item{isa}{an object of class \linkS4class{ISATab}, as retrieved by the
function \code{\link{readISAtab}}.}

\item{assay.filename}{a boolean indicating to show messages for the different
steps, if TRUE, or not to show them, if FALSE.}

\item{...}{extra arguments that can be pass down to the xcmsSet function from
the xcms package.}
}
\value{
the xcmsSet object built from the assay metadata, considering all the
        factors defined, and data files.
}
\description{
\code{processAssayXcmsSet.1factor} retrieves an xcmsSet object given an object
of class \linkS4class{ISATab} and one of its assay filenames.
}
\examples{
faahkoISA <- readISAtab(find.package("faahKO"))
assay.filename <- faahkoISA["assay.filenames"][[1]]
xset <- processAssayXcmsSet(faahkoISA, assay.filename)

}
\seealso{
\code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
