\name{processAssayXcmsSet}
\alias{processAssayXcmsSet}
\docType{data}
\title{ Builds an XCMS Set object given a mass spectrometry assay }
\description{
  \code{processAssayXcmsSet} reads a ISA-tab file.
}
\usage{
processAssayXcmsSet(path = getwd(),verbose=FALSE)
}
\arguments{
  \item{path}{ the name of the directory containing ISAtab files. The
    default is the working directory.}
  \item{verbose}{ a boolean indicating to show messages for the different steps, if TRUE, or not to show them, if FALSE}
  }
\value{
  
}
\author{
  Steffen Neumann, Alejandra Gonzalez-Beltran / 
  Maintainer: Audrey Kauffmann, Gabriella Rustici, Philippe Rocca-Serra, <isatools@googlegroups.com>  
}
\seealso{ \code{processAssayXcmsSet} }

\examples{
   
}
