% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risa.R
\docType{_PACKAGE}
\name{risa}
\alias{risa}
\title{risa: An R package for species and habitat risk assessment}
\description{
The package risa is a comprehensive R toolkit for mapping and quantifying ecological risk for one or more species (or habitats). It provides utilities for delineating areas of interest, managing coordinate reference systems, computing kernel-density classifications, identifying overlapping hotspots, estimating exposure and consequence components of risk assessment, and creating thematic risk maps. Its core functions are based on InVEST modules for habitat and species risk assessment that transform raw occurrence and stressor data into standardized risk indices. Outputs may be returned as both raster and polygon layers and easily exported as GeoTIFFs or Shapefiles.
}
