% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{scale_vars}
\alias{scale_vars}
\title{Rescale a SpatRaster to a New Numeric Range}
\usage{
scale_vars(x, new_range, quiet = TRUE)
}
\arguments{
\item{x}{A `terra::SpatRaster` object containing the values to be rescaled.}

\item{new_range}{A numeric vector of length 2 giving the new minimum and
maximum values for rescaling (`c(new_min, new_max)`).}
}
\value{
A `terra::SpatRaster` object with values rescaled to the specified
  range.
}
\description{
Linearly rescales the values of a `terra::SpatRaster` to a new user-defined
numeric range. The output is returned as a `SpatRaster` with values rounded
to four decimal places.
}
\details{
The function performs linear rescaling using:
\deqn{
  x' = a + \frac{(x - \min(x)) (b - a)}{\max(x) - \min(x)}
}
where \eqn{a} and \eqn{b} are the user-defined minimum and maximum.

If the raster has no variation (`min == max`), the function throws an error.
}
\examples{
\dontrun{
library(terra)
r <- rast(matrix(runif(100), 10, 10))
r2 <- scale_vars(r, c(0, 1))
}

}
