% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa-biocViews.R
\name{suggestBiocPackage}
\alias{suggestBiocPackage}
\title{Suggest Bioconductor pacakges.}
\usage{
suggestBiocPackage(isa, bioc.version = bioconductor.version)
}
\arguments{
\item{isa}{an object of class \linkS4class{ISATab}.}

\item{bioc.version}{The version number of Bioconductor to be considered as
the source of the pacakges for suggestions.}
}
\description{
Suggest a list of packages available in Bioconductor, which may be useful for
processing the assay according to its technology and measurement types.
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
