% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{updateAssayMetadata}
\alias{updateAssayMetadata}
\title{Update metadata into the assay file.}
\usage{
updateAssayMetadata(isa, assay.filename, col.name, values)
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}

\item{assay.filename}{A character vector with the filename of an assay.}

\item{col.name}{A character vector with the name of the column of the assay
file to be modified.}

\item{values}{A character vector with the values to be added to the column of
the assay file: it could be a single value, and in this case the
value is repeated across the column, or it could be a list of
values (whose length must match the number of rows of the assay
file).}
}
\value{
An updated object from the \code{\link{ISAtab-class}}.
}
\description{
\code{updataAssayMetadata}: Updates metadata into a particular assay file. See
an example of use for a metabolite dataset at \url{https://github.com/sneumann/mtbls2}.
}
\examples{
### This example shows how to add values to the column "Derived Spectral Data File",
### assuming that the results are stored in the file "faahkoDSDF.txt"
faahkoISA <- readISAtab(find.package("faahKO"))
assay.filename <- faahkoISA["assay.filenames"][[1]]
updateAssayMetadata(isa = faahkoISA,
                    assay.filename,
                    col.name = "Derived Spectral Data File",
                    values = "faahkoDSDF.txt")

}
\seealso{
\code{\link{readISAtab}}, \url{https://github.com/sneumann/mtbls2}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
