% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{write.assay.file}
\alias{write.assay.file}
\title{Write an Assay File from the ISA-Tab dataset.}
\usage{
write.assay.file(isa, assay.filename, path = getwd())
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}}

\item{assay.filename}{A character vector with the name of the assay file to
be written to disk.}

\item{path}{A character vector with the output path in which the assay file
is going to be written, by default the current working directory.}
}
\description{
The \code{write.assay.file} function writes a specific assay file from an
\code{\link{ISATab-class}} object into a file.
}
\examples{
### Example of writing the assay file of the faahKO ISA-Tab dataset into a
### temporary directory
faahkoISA <- readISAtab(path = find.package("faahKO"))
temp <- tempdir()
write.assay.file(isa = faahkoISA,
                 assay.filename = faahkoISA["assay.filenames"][[1]],
                 path = temp)

}
\seealso{
\code{\link{write.ISAtab}}, \code{\link{write.investigation.file}},
\code{\link{write.study.file}}, \code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
