% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{write.investigation.file}
\alias{write.investigation.file}
\title{Write the Investigation File from an ISA-Tab dataset.}
\usage{
write.investigation.file(isa, path = getwd())
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}

\item{path}{A character vector with the name of the directory to which the
investigation file is written. The default value is the current
working directory.}
}
\value{
The investigation file of an ISATab set of files.
}
\description{
\code{write.investigation.file} reads an \code{\link{ISATab-class}}
             object and writes the investigation part to a file.
}
\examples{
### Example of writing the investigation file of faahKO ISA-Tab dataset into
### a temporay directory
faahkoISA <- readISAtab(path = find.package("faahKO"))
temp <- tempdir()
write.investigation.file(isa = faahkoISA, path = temp)

}
\seealso{
\code{\link{write.ISAtab}}, \code{\link{write.study.file}},
\code{\link{write.assay.file}}, \code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
