% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{write.ISAtab}
\alias{write.ISAtab}
\title{Write an ISA-Tab dataset into files.}
\usage{
write.ISAtab(isa, path = getwd())
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}

\item{path}{A character vector with the name of the directory to which the
ISAtab files are written. The default value is the current working
directory.}
}
\value{
A set of ISATab files.
}
\description{
\code{write.ISAtab} writes a ISA-Tab dataset.
}
\examples{
### Example of writing the study file of faahKO ISA-Tab dataset into the temp director
data.dir <- system.file("extdata", package = "Risa")
isazip <- system.file("extdata", "faahKO-metadata.zip", package = "Risa")
faahkoISA <- readISAtab(path = file.path(data.dir, "faahKOISA"),
                        zipfile = isazip,
                        verbose = TRUE)
temp <- tempdir()
write.ISAtab(isa = faahkoISA,
             paht = temp)

}
\seealso{
\code{\link{write.investigation.file}}, \code{\link{write.study.file}},
\code{\link{write.assay.file}}, \code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
