% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risa.R
\name{write.study.file}
\alias{write.study.file}
\title{Write a Study File of an ISA-Tab dataset.}
\usage{
write.study.file(isa, study.filename, path = getwd())
}
\arguments{
\item{isa}{An object of the \code{\link{ISATab-class}}.}

\item{study.filename}{A character vector with the name of the study file to
be written to disk.}

\item{path}{A character vector with the output path in which the study file
is going to be written, by default the current working directory.}
}
\value{
A study file of an ISATab set of files.
}
\description{
The \code{write.study.file} function writes a specific Study file
             from an \code{\link{ISATab-class}} object into a file on disk.
}
\examples{
### Example of writing the study file of faahKO ISA-Tab dataset into a
### temporary directory
faahkoISA <- readISAtab(path = find.package("faahKO"))
temp <- tempdir()
write.study.file(isa = faahkoISA,
                 study.filename =  faahkoISA["study.filenames"][[1]],
                 path = temp)

}
\seealso{
\code{\link{write.ISAtab}}, \code{\link{write.investigation.file}},
\code{\link{write.assay.file}} \code{\link{readISAtab}}
}
\author{
Alejandra Gonzalez-Beltran (maintainer, ISA Team e-mail:\email{isatools@googlegroups.com})
}
