% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_maps.R
\name{export_maps}
\alias{export_maps}
\title{Export nested list of maps to files}
\usage{
export_maps(x, out_dir, path = character(), zip_export = FALSE)
}
\arguments{
\item{x}{A nested list of `SpatRaster`, `sf`, or further lists.}

\item{out_dir}{Character path to output directory.}

\item{path}{Internal use: character vector of path under recursion.}

\item{zip_export}{Logical; if `TRUE`, zip the top-level output.}
}
\value{
Invisibly returns `NULL`. Files are written to `out_dir`.
}
\description{
Recursively writes `SpatRaster` and `sf` layers, respectively as `.tif` and shapefiles, from a nested list to disk, optionally zipping.
}
\examples{
# Creating test data
spp_df <- rbind(data.frame(long = rnorm(80, 0, 10),
                           lat = rnorm(80, 0, 10), species = "sp1"),
                data.frame(long = rnorm(60, 0, 10),
                           lat = rnorm(60, 0, 10), species = "sp2"))
str_df <- rbind(data.frame(long = rnorm(100, 0, 5),
                           lat = rnorm(100, 0, 10), stressor = "trawling"),
                data.frame(long = rnorm(50, 0, 10),
                           lat = rnorm(100, 0, 5), stressor = "gillnet"))

# Create kernel maps of species and stressor distributions and overlap maps
risa_maps <- risa_prep(spp_df, str_df)

# Export vector and raster objects as shapefiles and tiff to external folder
export_maps(risa_maps, "my_folder")
}
