% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{[}
\alias{[}
\alias{ISATab-method}
\alias{ANY}
\alias{ANY-method}
\alias{[<-}
\alias{AssayTab-method}
\title{Extract methods}
\usage{
\S4method{[}{ISATab,ANY,ANY,ANY}(x, i)

\S4method{[}{ISATab,ANY,ANY,ANY}(x, i) <- value

\S4method{[}{AssayTab,ANY,ANY,ANY}(x, i)

\S4method{[}{AssayTab,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{S4 object of class \code{ISTab} or \code{AssayTab}}

\item{i}{character, name of a slot to extract or set}

\item{x}{S4 object of class \code{ISTab} or \code{AssayTab}}

\item{i}{character, name of a slot to extract or set}

\item{x}{S4 object of class \code{ISTab} or \code{AssayTab}}

\item{i}{character, name of a slot to extract or set}

\item{x}{S4 object of class \code{ISTab} or \code{AssayTab}}

\item{i}{character, name of a slot to extract or set}
}
\description{
extract slots from an object of class \linkS4class{ISATab}.
}
