% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{[}
\alias{[}
\alias{ISATab-method}
\alias{[}
\alias{[<-}
\alias{ISATab-method}
\alias{[}
\alias{AssayTab-method}
\alias{[}
\alias{[<-}
\alias{AssayTab-method}
\title{Extract methods}
\usage{
\S4method{[}{ISATab,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{ISATab,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[}{AssayTab,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{AssayTab,ANY,ANY,ANY}(x, i, j) <- value
}
\description{
extract slots from an object of class \linkS4class{ISATab}.

extract slots from an object of class \linkS4class{AssayTab}.
}
