% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_kernel.R
\name{get_class_kernel}
\alias{get_class_kernel}
\title{Kernel density estimation and reclassification wrapper}
\usage{
get_class_kernel(
  x,
  area = NULL,
  n_classes = 3,
  output_layer_type = "shp",
  radius = NULL,
  group_size = NULL
)
}
\arguments{
\item{x}{An `sf` object or data frame of points (lon/lat) or tibble.}

\item{area}{Bounding region (`bbox`, `sf`, or data frame) or `NULL` to auto-compute.}

\item{n_classes}{Integer number of reclassification bins.}

\item{output_layer_type}{Character; one of `"shp"`, `"raster"`, or `"both"`.}

\item{radius}{Numeric smoothing bandwidth; if `NULL`, set to 10 percent of mean pairwise distance.}

\item{group_size}{Optional column name for weights.}
}
\value{
An `sf` object, `SpatRaster`, or list containing both.
}
\description{
Computes KDE for point data, reclassifies into classes,
and returns a raster (`SpatRaster`), a vector (`sf`), or both.
The bandwidth of the KDE smoothing is based on the radius parameter.
If not informed, the radius is set as 10 percent of mean pairwise distance among points.
}
\examples{
# Creating test data
df <- data.frame(long = rnorm(120, 0, 10), lat = rnorm(120, 0, 10))

# Generating reclassified Kernel densities estimates (3 classes)
kde <- get_class_kernel(df)

# Plot KDE map
plot(kde)
}
