% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overlap_kernel.R
\name{get_overlap_kernel}
\alias{get_overlap_kernel}
\title{Compute overlap hotspots from two reclassified rasters}
\usage{
get_overlap_kernel(x, y, n_classes = 3, output_layer_type = "shp")
}
\arguments{
\item{x}{A reclassified `SpatRaster` of class codes.}

\item{y}{A reclassified `SpatRaster` with same CRS and class range.}

\item{n_classes}{Integer number of classes used in inputs.}

\item{output_layer_type}{Character; one of `"shp"`, `"raster"`, or `"both"`.}
}
\value{
An `sf` object, `SpatRaster`, or list containing both.
}
\description{
Multiplies two classed rasters and reclassifies the result into risk ratings.
}
\examples{
# Creating test data
species <- data.frame(long = rnorm(80, 0, 10), lat = rnorm(80, 0, 10))
stressor <- data.frame(long = rnorm(100, 0, 10), lat = rnorm(100, 0, 10))
kde_spe <- get_class_kernel(species,  output_layer_type = "raster")
kde_str <- get_class_kernel(stressor,  output_layer_type = "raster")
kde_spe <- terra::project(kde_spe, crs(kde_str), method = "bilinear")

# Creating reclassified overlap raster
overlap <- get_overlap_kernel(kde_spe, kde_str, output_layer_type = "raster")

# Plot KDE map
par(mfrow=c(1,3))
plot(kde_spe, main="Species distribution")
plot(kde_str, main="Stressor occurrence")
plot(overlap, main="Overlap")
}
