\name{getAnnotatedDataFrameAssay}
\alias{getAnnotatedDataFrameAssay}
\title{
Retrieves an AnnotatedDataFrameAssay.
}
\description{
Retrieves an AnnotatedDataFrameAssay.
}
\usage{
getAnnotatedDataFrameAssay(isa, assay.filename)
}
\arguments{
  \item{isa}{
An object of the \code{ISAtab-class}.
}
  \item{assay.filename}{
%A \code{character} .
}
}
%\details{
%}
\value{
An annotated data frame.
}
%\references{
%}
\author{
Alejandra Gonzalez-Beltran
}
%\note{
%}
\seealso{
\code{ISATab-class}
}
%\examples{
%}
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
