% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_tools.R
\name{guess_crs}
\alias{guess_crs}
\title{Guess coordinate reference system from geographic bounds}
\usage{
guess_crs(shp)
}
\arguments{
\item{shp}{An `sf` object or data frame with longitude/latitude in first two columns.}
}
\value{
Integer EPSG code (4326) or `NA` if bounds are invalid.
}
\description{
If coordinates fall within longitude [-180,180] and latitude [-90,90], returns EPSG:4326; else `NA`.
}
\examples{
# Create test data
coords <- data.frame(long = c(1,2,2,4), lat = c(4,4,2,2))

# Gussing coordinates
guess_crs(coords)

}
