% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tools.R
\name{merge_shp}
\alias{merge_shp}
\title{Merge a list of 'sf' objects into a data frame of coordinates}
\usage{
merge_shp(shp_list, group_size = NULL)
}
\arguments{
\item{shp_list}{A list of `sf` objects.}

\item{group_size}{Optional column name to include in output.}
}
\value{
A data.frame with columns `X`, `Y`, `group`, and optionally `group_size`.
}
\description{
Combines multiple vector objects of class `sf` into a single data frame with X, Y, and group labels.
}
\examples{
# Create test data
vec1 <- df_to_shp(data.frame(long = c(1,2,2,4), lat = c(4,4,2,2)))
vec2 <- df_to_shp(data.frame(long = c(2,5,4,6), lat = c(4,4,2,2)))
vec_list <- list(vec1, vec2)

# Convert vector list into data.frame
df <- merge_shp(vec_list)
df

}
