% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_matrix.R
\name{reclass_matrix}
\alias{reclass_matrix}
\title{Generate reclassification matrix for raster values}
\usage{
reclass_matrix(
  raster,
  n_classes = 3,
  exclude_lowest = TRUE,
  lowest_prop = 0.05
)
}
\arguments{
\item{raster}{A `terra::SpatRaster` object.}

\item{n_classes}{Integer 1–10 number of classes.}

\item{exclude_lowest}{Logical; if `TRUE`, values below 5 percent of max are mapped to `NA`.}

\item{lowest_prop}{Numeric; Defines the cutoff for the lowest proportion  (default `0.05`).}
}
\value{
A numeric matrix with columns `from`, `to`, and `class` for use with `terra::classify()`.
}
\description{
Builds a matrix mapping value ranges to discrete class codes,
optionally excluding lowest 5 percent.
}
\examples{
# Example
r <- terra::rast(nrows=5, ncols=5, vals=c(0.01,2,3,4,5))
reclass_matrix(r)
}
