\name{suggestBiocPackage}
\alias{suggestBiocPackage}
\title{
Suggest Bioconductor pacakges 
}
\description{
Suggest a list of packages available in Bioconductor, which may be useful for processing the assay according to its technology and measurement types.
}
\usage{
suggestBiocPackage(isa,bioc.version)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{isa}{ An \code{ISAtab-class} object. }
  \item{bioc.version}{ The version number of Bioconductor to be considered as the source of the pacakges for suggestions.}
}
%\details{
%}
%\value{
%}
%\references{
%}
\author{
Alejandra Gonzalez-Beltran
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
