\name{write.investigation.file}
\alias{write.investigation.file}
\docType{data}
\title{ Writes the Investigation File of an ISA-Tab dataset }
\description{
  \code{write.investigation.file} reads a ISA-tab file.
}
\usage{
write.investigation.file(path = getwd(),verbose=FALSE)
}
\arguments{
  \item{path}{ the name of the directory containing ISAtab files. The
    default is the working directory.}
  \item{verbose}{ a boolean indicating to show messages for the different steps, if TRUE, or not to show them, if FALSE}
  }
\value{
  
}
\author{
  Alejandra Gonzalez-Beltran / 
  Maintainer: Alejandra Gonzalez-Beltran, ISA Team <isatools@googlegroups.com>  
}
\seealso{ \code{write.investigation.file} }

\examples{
   
}