\name{write.isatab}
\alias{write.isatab}
\docType{data}
\title{ Write an ISA-Tab dataset }
\description{
  \code{write.isatab} reads a ISA-tab file.
}
\usage{
write.isatab(path = getwd(),verbose=FALSE)
}
\arguments{
  \item{path}{ the name of the directory containing ISAtab files. The
    default is the working directory.}
  \item{verbose}{ a boolean indicating to show messages for the different steps, if TRUE, or not to show them, if FALSE}
  }
\value{
  
}
\author{
  Alejandra Gonzalez-Beltran / 
  Maintainer: Alejandra Gonzalez-Beltran, ISA Team <isatools@googlegroups.com> 
}
\seealso{ \code{write.isatab} }

\examples{
   
}